requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
asyncpg==0.29.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.12.1
redis==5.0.1
passlib[argon2,bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
PyJWT==2.8.0
python-multipart==0.0.6
pydantic[email]==2.5.0
pydantic-settings==2.1.0
python-json-logger==2.0.7
httpx==0.25.2
python-dotenv==1.0.0
pyotp==2.9.0
qrcode[pil]==7.4.2

# Testing dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2
aiosqlite==0.19.0
common_security @ file:./packages/backend/common_security